+++ src/Manifold/Pipe.php
@@ -9,13 +9,10 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Systemic\Manifold;
 
-use DecodeLabs\Systemic\Manifold;
-use DecodeLabs\Systemic\ManifoldTrait;
+use DecodeLabs\Systemic\ManifoldAbstract;
 
-class Pipe implements Manifold
+class Pipe extends ManifoldAbstract
 {
-    use ManifoldTrait;
-
     /**
      * @return array<int, resource|array<string>>
      */